🔄 :: 코드 정렬

+++ src/main/java/com/example/entry/domain/feed/domain/repository/FeedRepository.java
@@ -3,5 +3,5 @@ package com.example.entry.domain.feed.domain.repository;
 import com.example.entry.domain.feed.domain.Feed;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FeedRepository extends CrudRepository<Feed, Long> {
+public interface FeedRepository extends CrudRepository<Feed, Integer> {
 }
